fix static_cast

+++ include/graph/graph.hpp
@@ -57,7 +57,7 @@ class Graph {
       throw std::out_of_range("i=j cant add edge");
     }
     for (int ind = 1; ind < static_cast<int>(arrayV_.size()) -
-                                static_cast<int>(layPrev.getID()) - 1;
+                                layPrev.getID() - 1;
          ind++)
       arrayV_[layPrev.getID() + ind]++;
     arrayE_.insert(arrayE_.begin() + arrayV_[layPrev.getID()], layNext.getID());
